Adds invokable receiver objects
